pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "rag-temporal-pinecone-docintel-ui"
version = "0.1.0"
description = "Production RAG platform with Temporal workflows, Pinecone, Azure Doc Intelligence, and Streamlit UI"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "sqlmodel>=0.0.22",
    "python-multipart>=0.0.9",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.5.0",
    "httpx>=0.27.0",
    "openai>=1.50.0",
    "anthropic>=0.34.0",
    "pinecone-client>=5.0.0",
    "langchain>=0.3.0",
    "langchain-core>=0.3.0",
    "langchain-text-splitters>=0.3.0",
    "temporalio>=1.7.0",
    "streamlit>=1.38.0",
    "tiktoken>=0.7.0",
    "numpy>=1.26.0",
    "azure-ai-formrecognizer>=3.3.0",
    "python-docx>=1.1.0",
    "PyPDF2>=3.0.0",
    "sentence-transformers>=3.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0",
    "httpx>=0.27.0",
    "ruff>=0.6.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
pythonpath = ["."]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.hatch.build.targets.wheel]
packages = ["app"]
